Implementação do feature de arquivos e o inicio da implementação da feature de pessoas

+++ people-manager-server/src/main/java/sajadvpm/feature/person/PersonRepository.java
@@ -2,5 +2,5 @@ package sajadvpm.feature.person;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PersonRepository  extends JpaRepository<Person, Long> {
+public interface PersonRepository  extends JpaRepository<Person, Integer> {
 }
